# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ebookstore Api",
    "description": "ebookstore Api",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/books/{id}": {
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Update a book by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Books"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Book updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Failed to update book"
          }
        }
      },
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get a book by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Book found",
            "schema": {
              "$ref": "#/definitions/Books"
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete a book by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Book deleted successfully"
          },
          "400": {
            "description": "Invalid ID format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Failed to delete book"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Books": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "The Silent Shore"
        },
        "author": {
          "type": "string",
          "example": "Elaine Morgan"
        },
        "stock": {
          "type": "number",
          "example": 3
        },
        "pages": {
          "type": "number",
          "example": 310
        },
        "description": {
          "type": "string",
          "example": "A gripping tale of survival and secrets on a desolate coastline."
        },
        "genre": {
          "type": "string",
          "example": "Mystery"
        },
        "tags": {
          "type": "array",
          "example": [
            "suspense",
            "thriller",
            "coastal"
          ],
          "items": {
            "type": "string"
          }
        },
        "price": {
          "type": "number",
          "example": 14.99
        }
      }
    },
...
  }
}